Accept, Rejectボタンで、disableを連動させた。
